+++ src/lib/services/fetch.js
@@ -2,8 +2,9 @@
 
 import type { Models } from 'flowTypes';
 
+import { models } from 'lib/db';
 import moment from 'moment';
-import { read } from 'lib/services/tokens';
+import { read, setValidity } from 'lib/services/tokens';
 import config from 'config';
 import axios from 'axios';
 
@@ -23,7 +24,7 @@ export function fetch ({ id, token, permissions }: Fetch) {
 
 const fetchBlock = {};
 
-export async function tryFetch (models: Models, id: number) {
+export async function tryFetch (mdls: Models, id: number) {
   const token = await read(models, { id });
   const now = moment();
   const time = moment(token.updatedAt).add(config.fetch.refetchTimeout, 'ms');
@@ -37,4 +38,18 @@ export async function tryFetch (models: Models, id: number) {
   }
 
   return;
+}
+
+export async function setTokenValidity (promise: Promise<*>, apiToken: string) {
+  try {
+    const response = await promise;
+    await setValidity(models, true, apiToken);
+    return response;
+  } catch (err) {
+    if (err.status === 400) {
+      await setValidity(models, false, apiToken);
+    }
+
+    throw err;
+  }
 }