+++ src/migration/scripts/2017-04-28-add-invalid-to-token.js
@@ -1,0 +1,7 @@
+const newFields = require('../lib/newFields');
+
+const fields = (Sequelize = {}) => ({
+  valid: Sequelize.BOOLEAN,
+});
+
+module.exports = newFields(fields, 'Gw2ApiToken');