+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+### Fixed
+- Pvp fetcher fetching when user doesn't have pvp permissions
+- Guilds of the day to be cached for the day
+
+### Changed
+- Fetch logging to be more granular
+
+### Added
+- Api token invalidator
+- Log when unhandled rejections/errors happen
 
 ## [3.1.0][] - 2017-04-28
 ### Added